refactor: move login out of resources
